refactor: extract Sidebar and FlashBanner from AdminLayout
- Move sidebar navigation logic to dedicated Sidebar component
- Extract FlashBanner into reusable component
- Simplify AdminLayout to compose components instead of inline implementation
- Improve code organization and reusability of admin UI components
